Blog page component breakdown update: First part of changes to break down blog page into smaller components. Main reason for this change is the design of the FeedItem popup will include parts of the Blog page in a different order.
This is the addition of the Review component. It will take in the user id of the profile that we are looking at as well as the current logged in user (this is for displaying a 'delete' function on the posts - whether or not anyone actually can is still determined in security rules) and an initial # of posts to display. The ability to see more posts is contained within this component